+++ docs/history.rst
@@ -24,6 +24,9 @@ Version 1.7.0 (Next version, not yet released)
 1. ``declarations.has_trivial_copy`` was defintevely removed.
    Please use ``declarations.has_copy_constructor``.
 
+2. Remove ``gccxml`` logger from logger class (was deprecated).
+   Please use ``cxx_parser`` instead.
+
 Version 1.6.2
 -------------